+++ agents/generate/page-data/compose-pages-data.mjs
@@ -294,12 +294,12 @@ function createCompositeInstance(section, component, componentLibrary, instanceI
 
   // Generate complete instances for each relatedComponent
   const relatedInstances = relatedComponents
-    .map(({ componentId, fieldCombinations: _fieldCombinations }, index) => {
+    .map(({ componentId, fieldCombinations: originalFieldCombinations }, index) => {
       log(`      🔍 Processing Related component ${index + 1}: ${componentId}`);
 
-      let fieldCombinations = _.cloneDeep(_fieldCombinations);
+      let fieldCombinations = _.cloneDeep(originalFieldCombinations);
       // first try to pick section by fieldCombinations
-      let childrenSection = _.pick(section, _fieldCombinations);
+      let childrenSection = _.pick(section, originalFieldCombinations);
 
       // check if it is a single list item
       const isSingleListItem =
@@ -312,21 +312,21 @@ function createCompositeInstance(section, component, componentLibrary, instanceI
       if (isSingleListItem) {
         // list get inside keys to match
         fieldCombinations = _.cloneDeep(
-          extractContentFields(_.get(section, _fieldCombinations[0])),
+          extractContentFields(_.get(section, originalFieldCombinations[0])),
         );
-        childrenSection = _.get(section, _fieldCombinations[0]);
+        childrenSection = _.get(section, originalFieldCombinations[0]);
       } else if (isSingleObjectItem) {
         childrenSection = {
           ...childrenSection,
-          ..._.get(section, _fieldCombinations[0]?.split(".")?.[0]),
+          ..._.get(section, originalFieldCombinations[0]?.split(".")?.[0]),
         };
       }
 
       // Find corresponding component in component library
       const relatedComponent = componentLibrary.find((comp) => {
-        if (componentId && comp.componentId === componentId) {
-          return true;
-        }
+        // if (componentId && comp.componentId === componentId) {
+        //   return true;
+        // }
 
         // fallback
         return _.isEqual(comp.fieldCombinations?.sort(), fieldCombinations?.sort());
@@ -334,7 +334,7 @@ function createCompositeInstance(section, component, componentLibrary, instanceI
 
       if (!relatedComponent) {
         log(
-          `      ❌ Component not found: ${componentId || "Unknown ID"} ${JSON.stringify(fieldCombinations)}`,
+          `      ❌ Component not found: ${componentId || "Unknown ID"} ${JSON.stringify(originalFieldCombinations)}`,
         );
         return null;
       }
@@ -346,12 +346,12 @@ function createCompositeInstance(section, component, componentLibrary, instanceI
         childrenSection,
         relatedComponent,
         componentLibrary,
-        index, // Pass the same section index
+        `${instanceId}-${index}`, // Pass the same section index
       );
 
       return {
         originalComponentId: relatedComponent.componentId,
-        originalGridSettingsKey: getChildFieldCombinationsKey(fieldCombinations),
+        originalGridSettingsKey: getChildFieldCombinationsKey(originalFieldCombinations),
         instanceId: childInstance.id,
         instance: childInstance,
         childrenSection,
@@ -407,8 +407,8 @@ function createComponentInstance(section, component, componentLibrary = [], sect
     sectionIndex,
     // Use hash of key fields to ensure same content generates same ID
     keys: Object.keys(section).sort(),
-    section,
   });
+
   const instanceId = generateDeterministicId(contentHash);
   log(`    🔧 Generated component instance ID: ${instanceId}`);
 
@@ -507,7 +507,6 @@ export default async function composePagesData(input) {
     componentLibrary,
     locale,
     translateLanguages = [],
-    path,
     pagesDir,
     outputDir,
     moreContentsComponentList,
@@ -527,7 +526,7 @@ export default async function composePagesData(input) {
     moreContentsComponentMap[comp.content.id] = comp;
   });
 
-  log(`🔧 Starting to compose Pages Kit YAML: ${path}`);
+  log(`🔧 Starting to compose Pages Kit YAML: ${pagesDir}`);
   log(`🧩 Component library count: ${componentLibrary?.length || 0}`);
   log(`🌐 Locale: ${locale}`);
   log(`📁 Output directory: ${pagesDir}`);
@@ -540,7 +539,7 @@ export default async function composePagesData(input) {
   const fileDataMap = new Map();
 
   if (middleFormatFiles && Array.isArray(middleFormatFiles)) {
-    log(`📄 Middle format files count: ${middleFormatFiles.length}`);
+    log(`📄 Page detail count: ${middleFormatFiles.length}`);
 
     // Build processing list containing all language files
     const filesToProcess = [];